twitter_streaming(GCP)/requirements.txt
google-cloud-pubsub==1.6.0
google-auth==2.3.3
tweepy==3.8.0